// UsmUI/src/models/allstocks.ts
export class allstocks{
    symbol?: string;
    securityName?: string;
    isinNo?: string;
    sector?: string;
    industry?: string;
    exchange?:string;
    currency?: string;
    securityCode?: string;
    gics?:string;
    price?:string;
    country?:string;
    close?:string;
    last?:string;
    isEdit?:false
}



//     isin="securityCode"?: "500014",
//     "symbol"?: "UTIQUE",
//     "securityName"?: "Utique Enterprises Ltd",
//     "sector"?: "Commodities",
//     "industry"?: "Metals & Mining",
//     "isinNo"?: "INE096A01010"

//     symbol="securityCode"?: "500010",
//     "symbol"?: "HDFC",
//     "securityName"?: "HOUSING DEVELOPMENT FINANCE CORP.LTD.",
//     "isinNo"?: "INE001A01036",
//     "sector"?: "Financial Services",
//     "industry"?: "Financial Services",
//     "exchange"?: "NSE",
//     "currency"?: "INR"

//     sector=
//     "securityCode"?: "500012",
//     "symbol"?: "ANDHRAPET",
//     "securityName"?: "ANDHRA PETROCHEMICALS LTD.",
//     "isinNo"?: "INE714B01016",
//     "sector"?: "Commodities",
//     "industry"?: "Chemicals",
//     "exchange"?: "NSE",
//     "currency"?: "INR"

//     security=
//     "securityCode"?: "500014",
//     "symbol"?: "UTIQUE",
//     "securityName"?: "Utique Enterprises Ltd",
//     "sector"?: "Commodities",
//     "industry"?: "Metals & Mining",
//     "isinNo"?: "INE096A01010"


// industry=
// "securityCode"?: "500002",
// "symbol"?: "ABB",
// "securityName"?: "ABB India Limited",
// "sector"?: "Industrials",
// "industry"?: "Capital Goods",
// "isinNo"?: "INE117A01022"